Somewhat enhance design

+++ LocationTest/LocationTest/Pages/MainPage.cs
@@ -12,6 +12,8 @@ namespace LocationTest.Pages
     {
         private readonly StackLayout ConnectedDevicesLayout = new StackLayout();
 
+        private readonly Button UploadButton;
+
         private readonly IBluetoothService BluetoothService;
 
         private readonly ILambdaFunctionDataService LambdaFunctionDataService;
@@ -47,6 +49,13 @@ namespace LocationTest.Pages
                 VerticalOptions = LayoutOptions.Center
             };
 
+            Label graphLabel = new Label
+            {
+                Text = "Show Graphs",
+                FontSize = 14,
+                Margin = new Thickness(10, 30, 0, 5),
+                VerticalOptions = LayoutOptions.Center
+            };
 
             Button buttonConnectBle = new Button
             {
@@ -59,21 +68,33 @@ namespace LocationTest.Pages
             };
             buttonConnectBle.Clicked += this.OnConnect;
 
-            Button buttonPlot = new Button
+
+            this.UploadButton = new Button
             {
-                Text = "Show thorax",
+                Text = "Upload data",
                 HorizontalOptions = LayoutOptions.Center,
                 IsEnabled = true,
                 IsVisible = true,
                 Margin = new Thickness(00, 10, 0, 0),
                 Padding = new Thickness(32, 5)
             };
-            buttonPlot.Clicked += (_, __) => this.OnButtonClicked(new ThoraxZoomModel());
+            this.UploadButton.Clicked += this.ButtonUpload_Clicked;
 
+
+            Button buttonThorax = new Button
+            {
+                Text = "Thorax Graph",
+                HorizontalOptions = LayoutOptions.CenterAndExpand,
+                IsEnabled = true,
+                IsVisible = true,
+                Margin = new Thickness(00, 10, 0, 0),
+                Padding = new Thickness(32, 5)
+            };
+            buttonThorax.Clicked += (_, __) => this.OnButtonClicked(new ThoraxZoomModel());
             Button buttonUV = new Button
             {
-                Text = "Show UV",
-                HorizontalOptions = LayoutOptions.Center,
+                Text = "UV Graph",
+                HorizontalOptions = LayoutOptions.CenterAndExpand,
                 IsEnabled = true,
                 IsVisible = true,
                 Margin = new Thickness(00, 10, 0, 0),
@@ -83,8 +104,8 @@ namespace LocationTest.Pages
 
             Button buttonPM = new Button
             {
-                Text = "Show PM",
-                HorizontalOptions = LayoutOptions.Center,
+                Text = "PM Graph",
+                HorizontalOptions = LayoutOptions.CenterAndExpand,
                 IsEnabled = true,
                 IsVisible = true,
                 Margin = new Thickness(00, 10, 0, 0),
@@ -93,15 +114,15 @@ namespace LocationTest.Pages
             buttonPM.Clicked += (_, __) => this.OnButtonClicked(new PMZoomModel());
             Button buttonMovement = new Button
             {
-                Text = "Show movement",
-                HorizontalOptions = LayoutOptions.Center,
+                Text = "Movement Graph",
+                HorizontalOptions = LayoutOptions.CenterAndExpand,
                 IsEnabled = true,
                 IsVisible = true,
                 Margin = new Thickness(00, 10, 0, 0),
                 Padding = new Thickness(32, 5)
             };
             buttonMovement.Clicked += (_, __) => this.OnButtonClicked(new MovementDataZoomModel());
-            var buttonGrid = new Grid
+            Grid buttonGrid = new Grid
             {
                 ColumnDefinitions = new ColumnDefinitionCollection
                 {
@@ -110,32 +131,37 @@ namespace LocationTest.Pages
                 }
             };
             buttonGrid.Children.Add(buttonConnectBle);
-            buttonGrid.Children.Add(buttonPlot);
-            Grid.SetColumn(buttonPlot, 1);
+            buttonGrid.Children.Add(this.UploadButton);
+            Grid.SetColumn(this.UploadButton, 1);
 
-            Button buttonUpload = new Button
+            Grid plotGrid = new Grid
             {
-                Text = "Upload data",
-                HorizontalOptions = LayoutOptions.Center,
-                IsEnabled = true,
-                IsVisible = true,
-                Margin = new Thickness(00, 10, 0, 0),
-                Padding = new Thickness(32, 5)
+                ColumnDefinitions = new ColumnDefinitionCollection
+                {
+                    new ColumnDefinition { Width = GridLength.Star },
+                    new ColumnDefinition { Width = GridLength.Star }
+                }
             };
-            buttonUpload.Clicked += ButtonUpload_Clicked;
+            plotGrid.Children.Add(buttonThorax);
+            plotGrid.Children.Add(buttonMovement);
+            plotGrid.Children.Add(buttonPM);
+            plotGrid.Children.Add(buttonUV);
+            Grid.SetRow(buttonPM, 1);
+            Grid.SetRow(buttonUV, 1);
+            Grid.SetColumn(buttonMovement, 1);
+            Grid.SetColumn(buttonUV, 1);
+
             this.Padding = new Thickness(5, Device.RuntimePlatform == Device.iOS ? 20 : 0, 5, 0);
             this.Content = new ScrollView
             {
                 Content = new StackLayout()
                 {
                     Children = {
-                        buttonUpload,
-                        buttonMovement,
-                        buttonPM,
-                        buttonUV,
                         buttonGrid,
                         connectedDevicesLabel,
-                        this.ConnectedDevicesLayout
+                        this.ConnectedDevicesLayout,
+                        graphLabel,
+                        plotGrid
                     }
                 }
             };
@@ -143,17 +169,22 @@ namespace LocationTest.Pages
 
         private async void ButtonUpload_Clicked(object sender, EventArgs e)
         {
-            
+            this.UploadButton.IsEnabled = false;
+            this.UploadButton.Text = "Uploading...";
+
             try
             {
                 string from = Path.Combine("storage", "emulated", "0", "Android", "data", "com.copd.COPDMonitor.Android", "files", "dataE24DDB60C06B.csv");
                 string to = Path.Combine("storage", "emulated", "0", "Android", "data", "com.copd.COPDMonitor.Android", "files", "uploadE24DDB60C06B.csv");
-                File.Copy(from, to);
-            File.Delete(from);
-           
-                await this.LambdaFunctionDataService.PostData(this.LoginResult.AccessToken, to);
+                if (File.Exists(from))
+                {
+                    File.Copy(from, to);
+                    File.Delete(from);
 
-                System.IO.File.Delete(to);
+                    await this.LambdaFunctionDataService.PostData(this.LoginResult.AccessToken, to);
+
+                    File.Delete(to);
+                }
             }
             catch (Exception exept)
             {
@@ -165,18 +196,24 @@ namespace LocationTest.Pages
             {
                 string from = Path.Combine("storage", "emulated", "0", "Android", "data", "com.copd.COPDMonitor.Android", "files", "dataCA81BA4BDC02.csv");
                 string to = Path.Combine("storage", "emulated", "0", "Android", "data", "com.copd.COPDMonitor.Android", "files", "uploadCA81BA4BDC02.csv");
-                File.Copy(from, to);
-                File.Delete(from);
+                if (File.Exists(from))
+                {
+                    File.Copy(from, to);
+                    File.Delete(from);
 
-                await this.LambdaFunctionDataService.PostData(this.LoginResult.AccessToken, to);
+                    await this.LambdaFunctionDataService.PostData(this.LoginResult.AccessToken, to);
 
-                System.IO.File.Delete(to);
+                    File.Delete(to);
+                }
             }
             catch (Exception exept)
             {
                 //De button is nu helemaal waardeloos, ik kijk er later na
                 Console.WriteLine(exept);
             }
+
+            this.UploadButton.IsEnabled = true;
+            this.UploadButton.Text = "Upload data";
         }
 
         private async void OnConnect(object sender, EventArgs e)
@@ -186,7 +223,7 @@ namespace LocationTest.Pages
 
         private void OnButtonClicked(IGraphZoomModel model)
         {
-            this.Navigation.PushAsync(new LinePlotView(this.LoginResult,model));
+            this.Navigation.PushAsync(new LinePlotView(this.LoginResult, model));
         }
 
         private void UpdateDevices()
